@@ -15,19 +15,16 @@
   <name>Sprint Boot Backend</name>
   <description>Realworld backed implementation</description>
   <properties>
-    <java.version>11</java.version>
+    <java.version>21</java.version>
     <project-base-package>io.realworld.backend</project-base-package>
     <spring.boot.mainclass>io.realworld.backend.BackendApplication</spring.boot.mainclass>
     <!-- Needed for animal-sniffer-maven-plugin version 1.19 (version 1.20 is fixed). -->
     <animal.sniffer.skip>true</animal.sniffer.skip>
     <!-- Dependency versions: -->
-    <jjwt.version>0.9.1</jjwt.version>
-    <checkerframework.version>3.12.0</checkerframework.version>
-    <checker-guava.version>30.1.1-jre</checker-guava.version>
-    <guava-failureaccess.version>1.0.1</guava-failureaccess.version>
-    <!--Copied from openapi-generator-maven-plugin output: -->
-    <springfox-version>2.9.2</springfox-version>
-    <swagger-version>1.6.2</swagger-version>
+    <jjwt.version>0.13.0</jjwt.version>
+    <checkerframework.version>3.49.0</checkerframework.version>
+    <checker-guava.version>33.1.0.2-jre</checker-guava.version>
+    <guava-failureaccess.version>1.0.3</guava-failureaccess.version>
   </properties>
   <dependencies>
     <!-- Spring dependencies: -->
