@@ -1,9 +1,9 @@
 package io.realworld.backend.domain.aggregate.comment;
 
 import java.util.List;
-import org.springframework.data.repository.CrudRepository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommentRepository extends CrudRepository<Comment, Long> {
+public interface CommentRepository extends JpaRepository<Comment, Long> {
   List<Comment> findByArticleId(long articleId);
 
   void deleteByArticleId(long articleId);
