@@ -1,11 +1,11 @@
 package io.realworld.backend.domain.aggregate.favourite;
 
 import java.util.List;
+import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.CrudRepository;
 
 public interface ArticleFavouriteRepository
-    extends CrudRepository<ArticleFavourite, ArticleFavouriteId> {
+    extends JpaRepository<ArticleFavourite, ArticleFavouriteId> {
 
   class FavouriteCount {
     private final long articleId;
